🐛[Fix] ShowCard Border 겹침 버그 수정 ♻️[Refactor] ShowCard onClick 영역 포스터 영역으로 변경
